refactor: move authn funcs out of auth routes
